# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WashBuddy API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings/{bookingId}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking detail
      parameters:
        - $ref: "#/components/parameters/BookingId"
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking:
                    $ref: "#/components/schemas/BookingDetail"
                required: [booking]
...
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review for a completed booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
  /reviews/provider:
    get:
      operationId: getProviderReviews
      tags: [reviews]
      summary: Get reviews for provider locations
      parameters:
        - name: locationId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Provider reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderReviewsResponse"
  /reviews/admin:
    get:
      operationId: getAdminReviews
      tags: [reviews]
      summary: Get all reviews for admin moderation
      parameters:
        - name: flagged
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Admin reviews
          content:
            application/json:
              schema:
                type: object
                properties:
                  reviews:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminReview"
                required: [reviews]
  /reviews/aggregate/{locationId}:
    get:
      operationId: getReviewAggregate
      tags: [reviews]
      summary: Get rating aggregate for a location
      parameters:
        - $ref: "#/components/parameters/LocationId"
      responses:
        "200":
          description: Rating aggregate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewAggregate"
  /locations/{locationId}/reviews:
    get:
      operationId: getLocationReviews
      tags: [reviews]
      summary: Get paginated reviews for a location
      parameters:
        - $ref: "#/components/parameters/LocationId"
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Location reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationReviewsResponse"
  /reviews/{reviewId}/reply:
    post:
      operationId: replyToReview
      tags: [reviews]
      summary: Provider replies to a customer review
      parameters:
        - $ref: "#/components/parameters/ReviewId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reply:
                  type: string
              required: [reply]
      responses:
        "200":
          description: Reply added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{reviewId}:
    patch:
      operationId: editReview
      tags: [reviews]
      summary: Edit a review within 72 hours
      parameters:
        - $ref: "#/components/parameters/ReviewId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rating:
                  type: integer
                comment:
                  type: string
      responses:
        "200":
          description: Review updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{reviewId}/vote:
    post:
      operationId: voteOnReview
      tags: [reviews]
      summary: Vote a review as helpful or unhelpful
      parameters:
        - $ref: "#/components/parameters/ReviewId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isHelpful:
                  type: boolean
              required: [isHelpful]
      responses:
        "200":
          description: Vote recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  isHelpful:
                    type: boolean
                required: [id, isHelpful]
  /reviews/{reviewId}/hide:
    post:
      operationId: hideReview
      tags: [reviews]
      summary: Admin hides an inappropriate review
      parameters:
        - $ref: "#/components/parameters/ReviewId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
              required: [reason]
      responses:
        "200":
          description: Review hidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
components:
  parameters:
...
    LocationId:
      name: locationId
      in: path
      required: true
      schema:
        type: string
...
    BookingId:
      name: bookingId
      in: path
      required: true
      schema:
        type: string
    ReviewId:
      name: reviewId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    UserRole:
      type: object
      properties:
        role:
          type: string
        scope:
          type: string
        scopeId:
          type: ["string", "null"]
      required: [role, scope]
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        roles:
          type: array
          items:
            $ref: "#/components/schemas/UserRole"
      required: [id, email, firstName, lastName, roles]
...
    Provider:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        legalName:
          type: ["string", "null"]
        slug:
          type: string
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        billingMode:
          type: string
        payoutReady:
          type: boolean
        createdAt:
          type: string
      required: [id, name, slug, billingMode, payoutReady]
...
    Location:
      type: object
      properties:
        id:
          type: string
        providerId:
          type: string
        name:
          type: string
        addressLine1:
          type: string
        city:
          type: string
        stateCode:
          type: string
        postalCode:
          type: string
        timezone:
          type: string
        isVisible:
          type: boolean
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        responseSlaFutureMins:
          type: integer
        responseSlaUnder1hMins:
          type: integer
        provider:
          $ref: "#/components/schemas/Provider"
        operatingWindows:
          type: array
          items:
            $ref: "#/components/schemas/OperatingWindow"
      required: [id, providerId, name, addressLine1, city, stateCode, postalCode, timezone, isVisible]
...
    OperatingWindow:
      type: object
      properties:
        id:
          type: string
        dayOfWeek:
          type: integer
        openTime:
          type: string
        closeTime:
          type: string
      required: [id, dayOfWeek, openTime, closeTime]
...
    Service:
      type: object
      properties:
        id:
          type: string
        locationId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        durationMins:
          type: integer
        basePriceMinor:
          type: integer
        platformFeeMinor:
          type: integer
        currencyCode:
          type: string
        capacityPerSlot:
          type: integer
        leadTimeMins:
          type: integer
        requiresConfirmation:
          type: boolean
        isVisible:
          type: boolean
      required: [id, locationId, name, durationMins, basePriceMinor, platformFeeMinor, currencyCode, capacityPerSlot, isVisible]
...
    BookingDetail:
      type: object
      properties:
        id:
          type: string
        locationId:
          type: string
        serviceId:
          type: string
        customerId:
          type: string
        vehicleId:
          type: ["string", "null"]
        status:
          type: string
        serviceNameSnapshot:
          type: string
        serviceBasePriceMinor:
          type: integer
        platformFeeMinor:
          type: integer
        totalPriceMinor:
          type: integer
        currencyCode:
          type: string
        scheduledStartAtUtc:
          type: string
        scheduledEndAtUtc:
          type: string
        providerCheckedInAtUtc:
          type: ["string", "null"]
        serviceStartedAtUtc:
          type: ["string", "null"]
        serviceCompletedAtUtc:
          type: ["string", "null"]
        createdAt:
          type: string
        location:
          $ref: "#/components/schemas/Location"
        service:
          $ref: "#/components/schemas/Service"
        customer:
          $ref: "#/components/schemas/AuthUser"
        vehicle:
          type: ["object", "null"]
        statusHistory:
          type: array
          items:
            $ref: "#/components/schemas/StatusHistoryEntry"
      required: [id, locationId, serviceId, customerId, status, serviceNameSnapshot, totalPriceMinor, currencyCode, scheduledStartAtUtc, scheduledEndAtUtc]
    StatusHistoryEntry:
      type: object
      properties:
        id:
          type: string
        fromStatus:
          type: ["string", "null"]
        toStatus:
          type: string
        changedBy:
          type: string
        reason:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, toStatus, changedBy, createdAt]
...
    CreateReviewInput:
      type: object
      properties:
        bookingId:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
      required: [bookingId, rating]
    Review:
      type: object
      properties:
        id:
          type: string
        bookingId:
          type: string
        authorId:
          type: string
        subjectId:
          type: string
        locationId:
          type: ["string", "null"]
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        isEdited:
          type: boolean
        editedAt:
          type: ["string", "null"]
        isHidden:
          type: boolean
        hiddenReason:
          type: ["string", "null"]
        providerReply:
          type: ["string", "null"]
        providerReplyAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        author:
          type: object
          properties:
            id:
              type: string
            firstName:
              type: string
            lastName:
              type: string
      required: [id, bookingId, authorId, rating, createdAt]
    ReviewAggregate:
      type: object
      properties:
        averageRating:
          type: ["number", "null"]
        totalReviews:
          type: integer
        distribution:
          type: object
          additionalProperties:
            type: integer
        responseRate:
          type: integer
      required: [totalReviews, distribution, responseRate]
    LocationReviewItem:
      type: object
      properties:
        id:
          type: string
        bookingId:
          type: string
        authorId:
          type: string
        authorName:
          type: string
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        isEdited:
          type: boolean
        providerReply:
          type: ["string", "null"]
        providerReplyAt:
          type: ["string", "null"]
        helpfulCount:
          type: integer
        unhelpfulCount:
          type: integer
        createdAt:
          type: string
      required: [id, bookingId, authorId, authorName, rating, createdAt]
    LocationReviewsResponse:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/LocationReviewItem"
        total:
          type: integer
        hasMore:
          type: boolean
        nextCursor:
          type: ["string", "null"]
        aggregate:
          $ref: "#/components/schemas/ReviewAggregate"
      required: [reviews, total, hasMore, aggregate]
    ProviderReviewItem:
      type: object
      properties:
        id:
          type: string
        bookingId:
          type: string
        authorName:
          type: string
        locationId:
          type: ["string", "null"]
        locationName:
          type: ["string", "null"]
        serviceName:
          type: string
        scheduledAt:
          type: string
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        isEdited:
          type: boolean
        providerReply:
          type: ["string", "null"]
        providerReplyAt:
          type: ["string", "null"]
        helpfulCount:
          type: integer
        createdAt:
          type: string
      required: [id, bookingId, authorName, rating, createdAt]
    ProviderReviewsResponse:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/ProviderReviewItem"
        needsReply:
          type: integer
        total:
          type: integer
      required: [reviews, needsReply, total]
    AdminReview:
      type: object
      properties:
        id:
          type: string
        bookingId:
          type: string
        authorName:
          type: string
        locationId:
          type: ["string", "null"]
        locationName:
          type: ["string", "null"]
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        isHidden:
          type: boolean
        hiddenReason:
          type: ["string", "null"]
        providerReply:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, bookingId, authorName, rating, isHidden, createdAt]
...