# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WashBuddy API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: string;
  locationId: string;
  name: string;
  /** @nullable */
  description?: string | null;
  durationMins: number;
  basePriceMinor: number;
  platformFeeMinor: number;
  currencyCode: string;
  capacityPerSlot: number;
  leadTimeMins?: number;
  requiresConfirmation?: boolean;
  isVisible: boolean;
}